+++ src/App.jsx
@@ -8,6 +8,7 @@ import PartnerPage from './pages/PartnerPage'
 import PartnersAdminPage from './pages/PartnersAdminPage'
 import ProposalsPage from './pages/ProposalsPage'
 import ProposalWizardPage from './pages/ProposalWizardPage'
+import AssessmentPage from './pages/AssessmentPage'
 import ChangePasswordPage from './pages/ChangePasswordPage'
 import ResetPasswordPage from './pages/ResetPasswordPage'
 import AppLayout from './components/layout/AppLayout'
@@ -63,6 +64,7 @@ export default function App() {
         />
 
         <Route path="/reset-password" element={<ResetPasswordPage />} />
+        <Route path="/assessment" element={<AssessmentPage />} />
 
         <Route
           path="/change-password"